Removes map from the interface
After thinking this through, map would be really tricky to implement.
When `search` is run we score the values. If a map is performed the
score for each item may no longer be an accurate representation of the
new item since it has been transformed. We could reset all scores back
to 1 when `map` is performed, but then what's the value in map? And on
top of that we would break one of the laws of a Functor (mapping
identity over it would not return the same Searchable). Therefore I no
longer see map as a valuable part of this interface. If a user wants to
transform the results they would be better served to just `fold` and
move on with their lives.